GlobalExceptionHandler로 예외 처리 옮기기, Controller 기본 채워넣기
