+++ src/main/java/com/example/devSns/controllers/ReplyController.java
@@ -4,6 +4,7 @@ import com.example.devSns.dto.ReplyDTO;
 import com.example.devSns.dto.ReplyResponse;
 import com.example.devSns.services.ReplyService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -15,23 +16,31 @@ import java.util.List;
 public class ReplyController {
     private final ReplyService replyService;
 
-    @GetMapping("/{postId}/comments")
-    public ResponseEntity<List<ReplyResponse>> getComments(@PathVariable long postId) {
-        return null;
+    @GetMapping("/{postId}/replies")
+    public ResponseEntity<List<ReplyResponse>> getReplies(@PathVariable long postId) {
+        List<ReplyResponse> replies = replyService.replyGetAll(postId);
+        ResponseEntity<List<ReplyResponse>> responseEntity = new ResponseEntity<>(replies, HttpStatus.OK);
+        return responseEntity;
     }
 
     @PostMapping("/{postId}")
-    public ResponseEntity<List<ReplyResponse>> writeComment(@PathVariable long postId, long userId, ReplyDTO comment) {
-        return null;
+    public ResponseEntity<ReplyResponse> writeReply(@PathVariable long postId, long userId, ReplyDTO reply) {
+        ReplyResponse replyResponse = replyService.writeReply(postId, userId, reply);
+        ResponseEntity<ReplyResponse> responseEntity = new ResponseEntity<>(replyResponse, HttpStatus.OK);
+        return responseEntity;
     }
 
     @PatchMapping("/{postid}")
-    public ResponseEntity<ReplyResponse> updateComment(@PathVariable long postId, long userId, ReplyDTO comment) {
-        return null;
+    public ResponseEntity<ReplyResponse> updateReply(@PathVariable long postId, long userId, ReplyDTO reply) {
+        ReplyResponse replyResponse = replyService.updateReply(postId, userId, reply);
+        ResponseEntity<ReplyResponse> responseEntity = new ResponseEntity<>(replyResponse, HttpStatus.OK);
+        return responseEntity;
     }
 
     @DeleteMapping("/{postId}")
-    public ResponseEntity<String> deleteComment(@PathVariable long postId, long userId) {
-        return null;
+    public ResponseEntity<String> deleteReply(@PathVariable long postId, long userId) {
+        String deleteCheck = replyService.deleteReply(postId, userId);
+        ResponseEntity<String> responseEntity = new ResponseEntity<>(deleteCheck, HttpStatus.OK);
+        return responseEntity;
     }
 }
